fix(expandable_player): cast clamp result to double for Slider value
The clamp() method returns num regardless of argument types. Adding
.toDouble() ensures the Slider receives the expected double type.

+++ lib/widgets/expandable_player.dart
@@ -884,7 +884,7 @@ class ExpandablePlayerState extends State<ExpandablePlayer>
                                   trackShape: const RoundedRectSliderTrackShape(),
                                 ),
                                 child: Slider(
-                                  value: currentProgress.clamp(0.0, currentTrack.duration!.inSeconds.toDouble()),
+                                  value: currentProgress.clamp(0.0, currentTrack.duration!.inSeconds.toDouble()).toDouble(),
                                   max: currentTrack.duration!.inSeconds.toDouble(),
                                   onChanged: (value) => setState(() => _seekPosition = value),
                                   onChangeStart: (value) => setState(() => _seekPosition = value),
